backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
python-dotenv
pytest
google-genai
